update:LexicalAnalysis
Se realizo un analisis lexico para determinar que tokens pertenecen y cuales no
