--- texto.txt
@@ -1,2 +1,0 @@
-ada -11 1 = 12 untextosuper11212largo ifvariable if + /
-prueba 12 for while fortuna 34 1 = 35 ifforwhile
