+++ texts/texto.txt
@@ -1,0 +1,1 @@
+*id1+id2(id3*id4)
